requirements.txt
langchain
llama-index
sentence-transformers
PyMuPDF
faiss-cpu
openai
huggingface_hub
python-dotenv